# internal/app/swagger/swagger.json
{"swagger":"2.0","info":{"contact":{},"license":{}},"paths":{"/web/v1/demos":{"post":{"tags":["Demo"],"summary":"创建数据","operationId":"Demo.Create","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/Demo","title":"Demo","type":"object"}}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/Demo","title":"Demo","type":"object"}},"400":{"description":"{error:{code:0,message:无效的请求参数}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}}},"/web/v1/demos/{id}":{"get":{"tags":["Demo"],"summary":"查询指定数据","operationId":"Demo.Get","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"path","name":"id","description":"记录ID","required":true,"type":"string"}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/Demo","title":"Demo","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"404":{"description":"{error:{code:0,message:资源不存在}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}},"put":{"tags":["Demo"],"summary":"更新数据","operationId":"Demo.Update","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"path","name":"id","description":"记录ID","required":true,"type":"string"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/Demo","title":"Demo","type":"object"}}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/Demo","title":"Demo","type":"object"}},"400":{"description":"{error:{code:0,message:无效的请求参数}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}},"delete":{"tags":["Demo"],"summary":"删除数据","operationId":"Demo.Delete","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"path","name":"id","description":"记录ID","required":true,"type":"string"}],"responses":{"200":{"description":"{status:OK}","schema":{"$ref":"#/definitions/HTTPStatus","title":"HTTPStatus","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}}},"/web/v1/demos/{id}/disable":{"patch":{"tags":["Demo"],"summary":"禁用数据","operationId":"Demo.Disable","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"path","name":"id","description":"记录ID","required":true,"type":"string"}],"responses":{"200":{"description":"{status:OK}","schema":{"$ref":"#/definitions/HTTPStatus","title":"HTTPStatus","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}}},"/web/v1/demos/{id}/enable":{"patch":{"tags":["Demo"],"summary":"启用数据","operationId":"Demo.Enable","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"path","name":"id","description":"记录ID","required":true,"type":"string"}],"responses":{"200":{"description":"{status:OK}","schema":{"$ref":"#/definitions/HTTPStatus","title":"HTTPStatus","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}}},"/web/v1/demos?q=page":{"get":{"tags":["Demo"],"summary":"查询分页数据","operationId":"Demo.QueryPage","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"query","name":"current","description":"分页索引","required":true,"type":"integer","format":"int32","default":1},{"in":"query","name":"pageSize","description":"分页大小","required":true,"type":"integer","format":"int32","default":10},{"in":"query","name":"code","description":"编号","type":"string"},{"in":"query","name":"name","description":"名称","type":"string"},{"in":"query","name":"status","description":"状态(1:启用 2:停用)","type":"integer","format":"int32"}],"responses":{"200":{"description":"查询结果：{list:列表数据,pagination:{current:页索引,pageSize:页大小,total:总数量}}","schema":{"type":"array","items":{"$ref":"#/definitions/Demo","title":"Demo","type":"object"}}},"400":{"description":"{error:{code:0,message:未知的查询类型}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}}},"/web/v1/pub/login":{"post":{"tags":["登录"],"summary":"用户登录","operationId":"登录.Login","parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/LoginInfo","title":"LoginInfo","type":"object"}}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/LoginTokenInfo","title":"LoginTokenInfo","type":"object"}},"400":{"description":"{error:{code:400,message:无效的请求参数}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:500,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}}},"/web/v1/pub/login/captcha":{"get":{"tags":["登录"],"summary":"响应图形验证码","operationId":"登录.ResCaptcha","parameters":[{"in":"query","name":"id","description":"验证码ID","required":true,"type":"string"},{"in":"query","name":"reload","description":"重新加载","type":"string"}],"responses":{"200":{"description":"图形验证码","schema":{"type":"file"}},"400":{"description":"{error:{code:0,message:无效的请求参数}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}}},"/web/v1/pub/login/captchaid":{"get":{"tags":["登录"],"summary":"获取验证码信息","operationId":"登录.GetCaptcha","responses":{"200":{"description":"","schema":{"$ref":"#/definitions/LoginCaptcha","title":"LoginCaptcha","type":"object"}}}}},"/web/v1/users":{"post":{"tags":["用户"],"summary":"创建数据","operationId":"用户.Create","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/UserInfo","title":"UserInfo","type":"object"}}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/UserInfo","title":"UserInfo","type":"object"}},"400":{"description":"{error:{code:0,message:无效的请求参数}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}}}},"definitions":{"Demo":{"title":"Demo","required":["code","name","status"],"type":"object","properties":{"code":{"description":"编号","type":"string"},"created_at":{"description":"创建时间","type":"string","format":"date-time"},"creator":{"description":"创建者","type":"string"},"memo":{"description":"备注","type":"string"},"name":{"description":"名称","type":"string"},"record_id":{"description":"记录ID","type":"string"},"status":{"description":"状态(1:启用 2:停用)","type":"integer","format":"int32"}}},"HTTPError":{"title":"HTTPError","required":["error"],"type":"object","properties":{"error":{"$ref":"#/definitions/HTTPErrorItem","description":"错误项","type":"object"}}},"HTTPErrorItem":{"title":"HTTPErrorItem","required":["code","message","trace_id"],"type":"object","properties":{"code":{"description":"错误码","type":"integer","format":"int32"},"message":{"description":"错误信息","type":"string"},"trace_id":{"description":"追踪Id","default":"用于快速定位错误","type":"string"}}},"HTTPStatus":{"title":"HTTPStatus","required":["status"],"type":"object","properties":{"status":{"description":"状态(OK)","type":"string"}}},"LoginCaptcha":{"title":"LoginCaptcha","type":"object","properties":{"captcha_id":{"description":"验证码ID","type":"string"}}},"LoginInfo":{"title":"LoginInfo","required":["type"],"type":"object","properties":{"captcha_code":{"description":"验证码","type":"string"},"captcha_id":{"description":"验证码ID","type":"string"},"password":{"description":"密码","type":"string"},"type":{"description":"登录状态","type":"string"},"user_name":{"description":"用户名","type":"string"}}},"LoginTokenInfo":{"title":"LoginTokenInfo","required":["access_token","token_type","expires_at"],"type":"object","properties":{"access_token":{"description":"访问令牌","type":"string"},"expires_at":{"description":"令牌到期时间","type":"integer","format":"int64"},"token_type":{"description":"令牌类型","type":"string"}}},"UserInfo":{"title":"UserInfo","required":["user_name","password","mobile"],"type":"object","properties":{"age":{"description":"年龄","type":"string"},"city":{"description":" 城市","type":"string"},"country":{"description":" 国家","type":"string"},"email":{"description":" 用户邮箱","type":"string"},"head_img_url":{"description":"微信头像","type":"string"},"id_card":{"description":" 身份认证正面","type":"string"},"id_card_no":{"description":" 身份认证反面","type":"string"},"last_login_time":{"description":"最后登录时间","type":"string","format":"date-time"},"mobile":{"description":"手机号","type":"string"},"nick_name":{"description":"微信昵称","type":"string"},"opened_id":{"description":"微信opened_id","type":"string"},"password":{"description":"密码","type":"string"},"photo":{"description":"头像","type":"string"},"province":{"description":"省份","type":"string"},"real_status":{"description":" 用户认证状态 auth:认证","type":"string"},"record_id":{"description":"记录id","type":"string"},"sex":{"description":"性别","type":"integer","format":"int32"},"status":{"description":"用户状态 1.启用 2.禁用","type":"integer","format":"int32"},"union_id":{"description":" 微信唯一id","type":"string"},"user_name":{"description":"名称","type":"string"}}}},"tags":[{"name":"用户","description":"用户接口"},{"name":"Demo","description":"示例接口"},{"name":"登录","description":"登录接口"}]}